fix: hostobjects fixes
