+++ packages/react-native-audio-api/common/cpp/audioapi/AudioAPIModuleInstaller.h
@@ -7,8 +7,8 @@
 #include <audioapi/HostObjects/AudioContextHostObject.h>
 #include <audioapi/HostObjects/OfflineAudioContextHostObject.h>
 #include <audioapi/HostObjects/inputs/AudioRecorderHostObject.h>
-#include <audioapi/HostObjects/AudioDecoderHostObject.h>
-#include <audioapi/HostObjects/AudioStretcherHostObject.h>
+#include <audioapi/HostObjects/utils/AudioDecoderHostObject.h>
+#include <audioapi/HostObjects/utils/AudioStretcherHostObject.h>
 
 #include <audioapi/events/AudioEventHandlerRegistry.h>
 #include <audioapi/HostObjects/events/AudioEventHandlerRegistryHostObject.h>
