+++ packages/react-native-audio-api/common/cpp/audioapi/HostObjects/utils/AudioStretcherHostObject.h
@@ -1,0 +1,28 @@
+#pragma once
+
+#include <audioapi/HostObjects/sources/AudioBufferHostObject.h>
+#include <audioapi/core/utils/AudioStretcher.h>
+#include <audioapi/jsi/JsiPromise.h>
+
+#include <jsi/jsi.h>
+#include <memory>
+#include <string>
+#include <thread>
+#include <utility>
+
+namespace audioapi {
+using namespace facebook;
+
+class AudioStretcherHostObject : public JsiHostObject {
+ public:
+  explicit AudioStretcherHostObject(
+      jsi::Runtime *runtime,
+      const std::shared_ptr<react::CallInvoker> &callInvoker,
+      float sampleRate);
+  JSI_HOST_FUNCTION_DECL(changePlaybackSpeed);
+
+ private:
+  std::shared_ptr<AudioStretcher> stretcher_;
+  std::shared_ptr<PromiseVendor> promiseVendor_;
+};
+} // namespace audioapi
